robert-agent/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robert-conversational-agent"
version = "1.0.0"
description = "Natural Swedish conversational agent for call reception and routing"
requires-python = ">=3.9"
dependencies = [
    "livekit-agents[openai]~=1.2",
    "python-dotenv~=1.0",
    "pyyaml~=6.0",
    "aiohttp~=3.9",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "pytest",
    "pytest-asyncio"
]